api/requirements.txt
cent
fastapi
pydantic
uvicorn
pyjwt